backend/requirements.txt
fastapi==0.98.0
uvicorn==0.23.1
pydantic==1.10.7
SQLAlchemy==1.4.48
requests==2.31.0
PyJWT==2.7.0
fastapi-limiter==0.1.6
redis==5.0.0
